Improve tick clock and no more physics entities passing through walls

+++ Collision.pde
@@ -1,11 +1,11 @@
 class Collision {
-	int tick;
+	long tick;
 	PVector position;
 	PVector normal;
 	float force;
 	Entity[] entities;
 
-	Collision(int tick, PVector position, PVector normal, float force, Entity[] entities) {
+	Collision(long tick, PVector position, PVector normal, float force, Entity[] entities) {
 		this.tick = tick;
 		this.position = position;
 		this.normal = normal;
